[src] Share [I]NSCollectionLayoutVisibleItem implementation between AppKit and UIKit.
